[analyzer] Apply @visibleOutsideTemplate to enclosed elements of annotated element
Currently if a class method is annotated with @visibleOutsideTemplate,
only the method declaration can be referred outside template files, but
not the enclosed method parameters.
Meanwhile, a method parameter is not a valid target for
@visibleOutsideTemplate, because it's not a member of Class, Enum or
Mixin.
Update the logic so that enclosed elements of a @visibleOutsideTemplate
element will also be @visibleOutsideTemplate.

Change-Id: Ifcfaa47a154fe6e2b15fae07f3a30e8a543916a2
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/317846
Commit-Queue: Ludi Zhan <ludizhan@google.com>
Reviewed-by: Samuel Rawlins <srawlins@google.com>